Bring Avian up to date + Assets

+++ src/test/miner_tests.cpp
@@ -1,5 +1,5 @@
 // Copyright (c) 2011-2016 The Bitcoin Core developers
-// Copyright (c) 2017 The Raven Core developers
+// Copyright (c) 2017-2019 The Raven Core developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
@@ -19,7 +19,7 @@
 #include "util.h"
 #include "utilstrencodings.h"
 
-#include "test/test_raven.h"
+#include "test/test_avian.h"
 
 #include <memory>
 
@@ -252,7 +252,7 @@ BOOST_FIXTURE_TEST_SUITE(miner_tests, TestingSetup)
         tx.vin[0].prevout.hash = txFirst[2]->GetHash();
         tx.vout.resize(2);
         tx.vout[0].nValue = 5000000000LL - 100000000;
-        tx.vout[1].nValue = 100000000; // 1RVN output
+        tx.vout[1].nValue = 100000000; // 1AVN output
         uint256 hashFreeTx2 = tx.GetHash();
         mempool.addUnchecked(hashFreeTx2, entry.Fee(0).SpendsCoinbase(true).FromTx(tx));
 
@@ -638,7 +638,7 @@ BOOST_FIXTURE_TEST_SUITE(miner_tests, TestingSetup)
         // it into the template because we still check IsFinalTx in CreateNewBlock,
         // but relative locked txs will if inconsistently added to mempool.
         // For now these will still generate a valid template until BIP68 soft fork
-        BOOST_CHECK_EQUAL(pblocktemplate->block.vtx.size(), 3);
+        BOOST_CHECK_EQUAL(pblocktemplate->block.vtx.size(), (uint64_t)3);
         // However if we advance height by 1 and time by 512, all of them should be mined
         for (int i = 0; i < CBlockIndex::nMedianTimeSpan; i++)
             chainActive.Tip()->GetAncestor(chainActive.Tip()->nHeight - i)->nTime += 512; //Trick the MedianTimePast
@@ -646,7 +646,7 @@ BOOST_FIXTURE_TEST_SUITE(miner_tests, TestingSetup)
         SetMockTime(chainActive.Tip()->GetMedianTimePast() + 1);
 
         BOOST_CHECK(pblocktemplate = AssemblerForTest(chainparams).CreateNewBlock(scriptPubKey));
-        BOOST_CHECK_EQUAL(pblocktemplate->block.vtx.size(), 5);
+        BOOST_CHECK_EQUAL(pblocktemplate->block.vtx.size(), (uint64_t)5);
 
         chainActive.Tip()->nHeight--;
         SetMockTime(0);
